[clang-tidy][no-assign-int-to-ptr] Detect vulnerability CWE-587
CWE-587 vulnerability means that assigning a constant address other than
'NULL' or '0' to a pointer is considered a bad practice. Using a fixed
address is not portable, because this address will not be valid in all
environments or platforms.
Description could be found in MiscNoAssignIntToPtr.md,
docs/clang-tidy/checks/misc/no-assign-int-to-ptr.rst